+++ WHERETFWILLITEND/WHERETFWILLITEND/Graphics/Constants.h
@@ -33,6 +33,6 @@ struct PassConstants {
 	LightData lights[128];
 	shaderMaterialData mats[300];
 	float max_lights;
-	float current_mat;
+	int current_mat;
 	float pad2[2];
 };
